all: move modules to stdlib
