+++ examples/multi-context/main.go
@@ -14,7 +14,7 @@ import (
 
 	// This initializes gpython for runtime execution and is critical.
 	// It defines forward-declared symbols and registers native built-in modules, such as sys and time.
-	_ "github.com/go-python/gpython/modules"
+	_ "github.com/go-python/gpython/stdlib"
 
 	// This is the primary import for gpython.
 	// It contains all symbols needed to fully compile and run python.
@@ -129,7 +129,7 @@ func RunMultiPi(numWorkers, numTimes int) time.Duration {
 			}
 			workersRunning.Done()
 
-			// This drives modules being able to perform cleanup and release resources 
+			// This drives modules being able to perform cleanup and release resources
 			w.ctx.Close()
 		}()
 	}
